Major changes to bashrc
Changed the prompt to a 2 line prompt showing the current terminal
and the directory on the first line, leaving the prompt on the
second. Also added an optional prompt which changes color according
to the battery status.
Added support for colored man pages, and removed and fixed a few
old aliases